+++ libs/dot-components/src/lib/components/switch/Switch.stories.tsx
@@ -4,7 +4,7 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import { DotSwitch, SwitchProps } from './Switch';
 
 export default {
-  title: 'Experimental/Switch',
+  title: 'Components/Switch',
   component: DotSwitch,
   argTypes: {
     ariaLabel: { defaultValue: 'Accessibility for the win' },